bootstrap volume to create db closes #6
