+++ src/app/[orgId]/settings/share-links/CreateShareLinkForm.tsx
@@ -525,7 +525,7 @@ export default function CreateShareLinkForm({
                                         </div>
                                         <CollapsibleContent className="space-y-2">
                                             {directLink && (
-                                                <div className="space-y-1">
+                                                <div className="space-y-2">
                                                     <div className="mx-auto">
                                                         <CopyTextBox
                                                             text={directLink}
